// src/app/enums/cgol-patterns.ts
export enum CgolPatterns {
  Still_Life_Block = 'Still_Life_Block',
  Still_Life_BeeHive = 'Still_Life_BeeHive',
  Still_Life_Loaf = 'Still_Life_Loaf',
  Still_Life_Boat = 'Still_Life_Boat',
  Still_Life_Tub = 'Still_Life_Tub',
  Oscillator_Blinker_Horizontal = 'Oscillator_Blinker_Horizontal',
  Oscillator_Blinker_Vertical = 'Oscillator_Blinker_Vertical',
  Oscillators_Toad = 'Oscillators_Toad',
  Oscillators_Beacon = 'Oscillators_Beacon',
  Spaceship_Glider = 'Spaceship_Glider',
  Gun_CIS_Bee_Shuttle = 'Gun_CIS_Bee_Shuttle',
  Gun_Trans_Bee_Shuttle = 'Gun_Trans_Bee_Shuttle'
  // Spaceship_Light_Weight,
  // Spaceship_Middle_Weight,
  // Spaceship_Heavy_Weight
}